+++ testutil/simsx/common_test.go
