+++ testutil/simsx/reporter.go
