+++ testutil/simsx/reporter_test.go
